further refactor into wcs and transforms modules
