+++ romancal/orientation/_pointing.py
@@ -8,6 +8,8 @@ import numpy as np
 from astropy.table import Table
 from astropy.time import Time, TimeDelta
 
+from . import _lib as olib
+
 from ..lib.engdb.engdb_lib import EngDB_Value
 from ..lib.engdb.engdb_tools import engdb_service
 
@@ -16,16 +18,12 @@ __all__ = []
 # Setup logging
 logger = logging.getLogger(__name__)
 logger.addHandler(logging.NullHandler())
-DEBUG_FULL = logging.DEBUG - 1
-LOGLEVELS = [logging.INFO, logging.DEBUG, DEBUG_FULL]
+LOGLEVELS = [logging.INFO, logging.DEBUG, olib.DEBUG_FULL]
 
 # Mnemonics needed.
 COARSE_MNEMONICS_QUATERNION_ECI = [f"SCF_AC_SDR_QBJ_{idx + 1}" for idx in range(4)]
 COARSE_MNEMONICS_B2FGS_EST = [f"SCF_AC_EST_FGS_QBR_{idx + 1}" for idx in range(4)]
 COARSE_MNEMONICS = COARSE_MNEMONICS_QUATERNION_ECI + COARSE_MNEMONICS_B2FGS_EST
-
-# Potential errors generated by mnemonic retrival
-EXPECTED_ERRORS = (OSError, RuntimeError, ValueError)
 
 # Pointing container
 # Attributes are as follows. Except for the observation time, all values
@@ -111,7 +109,7 @@ def get_mnemonics(
         service_kwargs = dict()
     try:
         engdb = engdb_service(**service_kwargs)
-    except EXPECTED_ERRORS as exception:
+    except olib.EXPECTED_ERRORS as exception:
         raise ValueError(
             f"Cannot open engineering DB connection\nException: {exception}"
         ) from exception
@@ -133,7 +131,7 @@ def get_mnemonics(
                 include_obstime=True,
                 include_bracket_values=False,
             )
-        except EXPECTED_ERRORS as exception:
+        except olib.EXPECTED_ERRORS as exception:
             logger.warning("Cannot retrieve %s from engineering.", mnemonic)
             logger.debug("Exception %s", exception)
             continue
@@ -245,8 +243,8 @@ def get_pointing(
     )
     reduced = reduce_func(mnemonics)
 
-    logger.log(DEBUG_FULL, "Mnemonics found:")
-    logger.log(DEBUG_FULL, "%s", mnemonics)
+    logger.log(olib.DEBUG_FULL, "Mnemonics found:")
+    logger.log(olib.DEBUG_FULL, "%s", mnemonics)
     logger.info("Reduced set of pointings:")
     logger.info("%s", reduced)
